# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Auth Api",
        "version": "v1"
    },
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "List user",
                "operationId": "user_index",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Send request success",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 200,
                                    "data": [
                                        {
                                            "id": 1,
                                            "name": "..........."
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Login false",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 200,
                                    "data": {
                                        "current_page": 1,
                                        "data": [
                                            {
                                                "id": 1,
                                                "name": "Admin",
                                                "email": "admin@gmail.com",
                                                "email_verified_at": "2022-03-01T17:07:32.000000Z",
                                                "created_at": "2022-03-01T17:07:32.000000Z",
                                                "updated_at": "2022-03-01T17:07:32.000000Z"
                                            },
                                            {
                                                "id": 2,
                                                "name": "QAM",
                                                "email": "aqm@gmail.com",
                                                "email_verified_at": "2022-03-01T17:07:32.000000Z",
                                                "created_at": "2022-03-01T17:07:32.000000Z",
                                                "updated_at": "2022-03-01T17:07:32.000000Z"
                                            },
                                            {
                                                "id": 3,
                                                "name": "QAC",
                                                "email": "aqc@gmail.com",
                                                "email_verified_at": "2022-03-01T17:07:32.000000Z",
                                                "created_at": "2022-03-01T17:07:32.000000Z",
                                                "updated_at": "2022-03-01T17:07:32.000000Z"
                                            },
                                            {
                                                "id": 4,
                                                "name": "STAFF",
                                                "email": "staff@gmail.com",
                                                "email_verified_at": "2022-03-01T17:07:32.000000Z",
                                                "created_at": "2022-03-01T17:07:32.000000Z",
                                                "updated_at": "2022-03-01T17:07:32.000000Z"
                                            }
                                        ],
                                        "first_page_url": "http://127.0.0.1:8000/api/users?per_page=10&page=1",
                                        "from": 1,
                                        "last_page": 1,
                                        "last_page_url": "http://127.0.0.1:8000/api/users?per_page=10&page=1",
                                        "links": [
                                            {
                                                "url": null,
                                                "label": "&laquo; Previous",
                                                "active": false
                                            },
                                            {
                                                "url": "http://127.0.0.1:8000/api/users?per_page=10&page=1",
                                                "label": "1",
                                                "active": true
                                            },
                                            {
                                                "url": null,
                                                "label": "Next &raquo;",
                                                "active": false
                                            }
                                        ],
                                        "next_page_url": null,
                                        "path": "http://127.0.0.1:8000/api/users",
                                        "per_page": "10",
                                        "prev_page_url": null,
                                        "to": 4,
                                        "total": 4
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Add new user",
                "operationId": "user_create",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "name": {
                                        "format": "string"
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "name": "Nguyen Nam 2222",
                                "email": "namnt@gmail.com",
                                "password": "paSSWORD",
                                "birth": "2000-05-27",
                                "role": 2
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Send request success",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 200
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Detail User",
                "operationId": "user_show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Send request success",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 200,
                                    "data": {
                                        "status": 200,
                                        "profile": {
                                            "id": 1,
                                            "name": "Admin",
                                            "email": "admin@gmail.com",
                                            "email_verified_at": "2022-03-01T17:07:32.000000Z",
                                            "created_at": "2022-03-01T17:07:32.000000Z",
                                            "updated_at": "2022-03-01T17:07:32.000000Z",
                                            "roles": [
                                                {
                                                    "id": 4,
                                                    "name": "ADMIN",
                                                    "guard_name": "api",
                                                    "created_at": "2022-03-01T17:07:32.000000Z",
                                                    "updated_at": "2022-03-01T17:07:32.000000Z",
                                                    "pivot": {
                                                        "model_id": 1,
                                                        "role_id": 4,
                                                        "model_type": "App\\\\Models\\\\User"
                                                    }
                                                }
                                            ]
                                        },
                                        "roles": [
                                            "ADMIN"
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Login false",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 401,
                                    "message": "Username or password invalid"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update User",
                "operationId": "user_update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "name": {
                                        "format": "string"
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "name": "Nguyen Nam 2222",
                                "new_password": "new_paSSWORD",
                                "birth": "2000-05-27",
                                "role": 2
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Send request success",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 200
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access Deny permission",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 403,
                                    "message": "Access Deny permission"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete User",
                "operationId": "user_delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Send request success",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 200
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth": []
                    }
                ]
            }
        },
        "/api/roles": {
            "get": {
                "tags": [
                    "Role"
                ],
                "summary": "List Roles",
                "operationId": "user_roles",
                "responses": {
                    "200": {
                        "description": "Send request success",
                        "content": {
                            "application/json": {
                                "example": [
                                    {
                                        "id": 1,
                                        "name": "QAM",
                                        "guard_name": "api",
                                        "created_at": "2022-03-05T18:01:41.000000Z",
                                        "updated_at": "2022-03-05T18:01:41.000000Z"
                                    },
                                    {
                                        "id": 2,
                                        "name": "QAC",
                                        "guard_name": "api",
                                        "created_at": "2022-03-05T18:01:41.000000Z",
                                        "updated_at": "2022-03-05T18:01:41.000000Z"
                                    },
                                    {
                                        "id": 3,
                                        "name": "STAFF",
                                        "guard_name": "api",
                                        "created_at": "2022-03-05T18:01:41.000000Z",
                                        "updated_at": "2022-03-05T18:01:41.000000Z"
                                    },
                                    {
                                        "id": 4,
                                        "name": "ADMIN",
                                        "guard_name": "api",
                                        "created_at": "2022-03-05T18:01:41.000000Z",
                                        "updated_at": "2022-03-05T18:01:41.000000Z"
                                    }
                                ]
                            }
                        }
                    },
                    "401": {
                        "description": "Login false",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 403
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth": []
                    }
                ]
            }
        }
    },
    "security": [
        []
    ]
}